requirements.txt
aiogram==3.3.0
aiohttp==3.9.1
aiohttp-cors==0.7.0
python-dotenv==1.0.0
sqlalchemy==2.0.25
aiosqlite==0.19.0
Pillow==10.2.0
cryptography==41.0.7
